requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
argcomplete==2.0.0
asgiref==3.7.2
attrs==23.2.0
Beaker==1.12.1
beautifulsoup4==4.12.3
blivet==3.8.2
blivet-gui==2.5.0
Brlapi==0.8.5
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.3
cryptography==41.0.7
cupshelpers==1.0
dasbus==1.7
dbus-python==1.3.2
decouple==0.0.7
distlib==0.3.6
distro==1.8.0
Django==5.0.3
django-crispy-forms==2.1
django-jalali-date==1.1.0
django-widget-tweaks==1.5.0
dnf==4.19.0
fedora-third-party==0.10
file-magic==0.4.0
filelock==3.12.2
fros==1.1
frozenlist==1.4.1
humanize==3.13.1
idna==3.4
isodate==0.6.1
jalali_core==1.0.0
jdatetime==5.0.0
langtable==0.0.65
libcomps==0.1.20
libdnf==0.73.0
lxml==4.9.3
Mako==1.2.3
MarkupSafe==2.1.3
multidict==6.0.5
nftables==0.1
olefile==0.46
packaging==23.1
Paste==3.5.3
pexpect==4.8.0
pid==2.2.3
Pillow==10.0.1
platformdirs==3.9.1
ply==3.11
productmd==1.38
ptyprocess==0.7.0
pwquality==1.4.5
pycairo==1.25.1
pycparser==2.20
pycrypto==2.6.1
pycups==2.0.1
pycurl==7.45.2
pyenchant==3.2.2
PyGObject==3.46.0
pykickstart==3.48
pyOpenSSL==23.2.0
pyparted==3.13.0
PyQt5==5.15.10
PyQt5-sip==12.13.0
PySocks==1.7.1
python-augeas==1.1.0
python-dateutil==2.8.2
python-meh==0.51
pytz==2023.3.post1
pyudev==0.24.1
pyxdg==0.27
PyYAML==6.0.1
regex==2023.12.25
requests==2.28.2
requests-file==1.5.1
requests-ftp==0.3.1
requests-toolbelt==1.0.0
rpm==4.19.1.1
selinux @ file:///builddir/build/BUILD/libselinux-3.5/src
sepolicy @ file:///builddir/build/BUILD/selinux-3.5/python/sepolicy
setools==4.4.3
setuptools==67.7.2
simpleaudio==1.0.4
simpleline==1.9.0
six==1.16.0
sos==4.6.0
soupsieve==2.5
sqlparse==0.4.4
systemd-python==235
Tempita==0.5.2
urllib3==1.26.18
virtualenv==20.21.1
yarl==1.9.4
zeep==4.2.1
